Refactored REST Client
* Generalised and pull out REST Client from libgds API module
  to make way for impl other web services
